apps/userbot/pyproject.toml
[project]
name = "bb-userbot"
version = "0.0.0"
description = "Telethon userbot: multiplex MTProto sessions, QR-login, ingest."
requires-python = ">=3.12"
dependencies = [
    "telethon>=1.40,<2.0",
    "cryptg>=0.5.0",
    "asyncpg>=0.30.0",
    "cryptography>=44.0.0",
    "structlog>=25.1.0",
    "python-dotenv>=1.0.1",
    "orjson>=3.10.14",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.2",
    "mypy>=1.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bb_userbot"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
